Conflicts:
	visaApplication/src/main/java/com/mastek/visaApplication/entities/ApplicationForm.java
	visaApplication/src/main/java/com/mastek/visaApplication/entities/PersonalDetails.java
	visaApplication/src/main/java/com/mastek/visaApplication/services/VisaApplicationServices.java
	visaApplication/src/test/java/com/mastek/visaApplication/MongoTests.java
	visaApplication/src/test/java/com/mastek/visaApplication/VisaApplicationTests.java

+++ visaApplication/src/main/java/com/mastek/visaApplication/entities/PersonalDetails.java
@@ -114,9 +114,13 @@ public class PersonalDetails {
 	String NationalInsuranceNumber;
 	
 	//Dependants//
+
 	@FormParam("dependantPassportNo")
-	long dependantPassportNo;
+	int dependantPassportNo;
 	
+	
+
+//CREATE DEPENDANTS PASSPORT NUMBER AND PASSPORT DETAILS//
 	@FormParam("dependantFamilyName")
 	String dependantFamilyName;
 	
@@ -410,11 +414,11 @@ public class PersonalDetails {
 		NationalInsuranceNumber = nationalInsuranceNumber;
 	}
 	
-	public long getDependantPassportNo() {
+	public int getDependantPassportNo() {
 		return dependantPassportNo;
 	}
 
-	public void setDependantPassportNo(long dependantPassportNo) {
+	public void setDependantPassportNo(int dependantPassportNo) {
 		this.dependantPassportNo = dependantPassportNo;
 	}
